create playStateHistory and createLastStateEventProps functions for managing game state animations
