+++ src/js/td-scenes/battle/procedure/play-state-history/create-last-state-event-props.ts
@@ -1,0 +1,44 @@
+import { GameState } from "gbraver-burst-core";
+
+import { getMainTurnCount } from "../../../../custom-battle-events/get-main-turn-count";
+import { separatePlayers } from "../../../../custom-battle-events/separate-players";
+import { LastStateEventProps } from "../../custom-battle-event";
+import { BattleSceneProps } from "../../props";
+
+/**
+ * LastStateEventPropsを生成する
+ * @param props 戦闘シーンプロパティ
+ * @param gameStateHistory ゲームステートヒストリー
+ * @returns eventProps
+ */
+export function createLastStateEventProps(
+  props: Readonly<BattleSceneProps>,
+  gameStateHistory: GameState[],
+): LastStateEventProps {
+  const lastState = gameStateHistory.at(-1);
+  if (!lastState) {
+    throw new Error("gameStateHistory is empty");
+  }
+  const separatedPlayers = separatePlayers(props, lastState);
+  const player = separatedPlayers?.player ?? lastState.players[0];
+  const enemy = separatedPlayers?.enemy ?? lastState.players[1];
+  const playerMainTurnCount = getMainTurnCount({
+    stateHistory: props.stateHistory,
+    playerId: player.playerId,
+  });
+  const enemyMainTurnCount = getMainTurnCount({
+    stateHistory: props.stateHistory,
+    playerId: enemy.playerId,
+  });
+  const mainTurnCount = playerMainTurnCount + enemyMainTurnCount;
+  return {
+    ...props,
+    player,
+    playerMainTurnCount,
+    mainTurnCount,
+    enemy,
+    enemyMainTurnCount,
+    update: gameStateHistory,
+    lastState,
+  };
+}
